+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [3.1.0] - 2019-04-10
 ### Added
 - new 'assessPassword' method, executes validation but does not save password
 
@@ -71,7 +73,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial version of ID Broker API client.
 
-[Unreleased]: https://github.com/silinternational/idp-id-broker-php-client/compare/3.0.0...HEAD
+[Unreleased]: https://github.com/silinternational/idp-id-broker-php-client/compare/3.1.0...HEAD
+[3.1.0]: https://github.com/silinternational/idp-id-broker-php-client/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/silinternational/idp-id-broker-php-client/compare/2.6.0...3.0.0
 [2.6.0]: https://github.com/silinternational/idp-id-broker-php-client/compare/2.5.1...2.6.0
 [2.5.1]: https://github.com/silinternational/idp-id-broker-php-client/compare/2.5.0...2.5.1